LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/infrastructure/urls/resolving-urls/query-encoding/windows-1252_include=workers-expected.txt


PASS Worker constructor
FAIL SharedWorker constructor Shared workers are not supported
PASS importScripts() in a dedicated worker
PASS Worker() in a dedicated worker
FAIL SharedWorker() in a dedicated worker assert_equals: expected "%C3%A5" but got "ReferenceError: Can't find variable: SharedWorker"
FAIL importScripts() in a shared worker Shared workers are not supported
FAIL Worker() in a shared worker Shared workers are not supported
FAIL SharedWorker() in a shared worker Shared workers are not supported
